requirements.txt
aiogram==3.13.1
asyncpg==0.30.0
aiosqlite==0.21.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
loguru==0.7.2
redis==5.1.1
python-dotenv==1.0.1
httpx==0.27.2
openpyxl==3.1.5
aiofiles==24.1.0
openai-whisper==20250625
ffmpeg-python==0.2.0

